We have to merge all of the commits to clean things up. After 6 mos or so any user who will look into it will go crazy We lost trace of how we do the commits but this is the best we could.
Finish Login Logic

Merging All of the Commits

We have to merge all of the commits to clean things up.
After 6 mos or so any user who will look into it will go crazy
We lost trace of how we do the commits but this is the best we could.

Finish Services API